+++ storm-ktor/src/test/resources/schema.sql
@@ -1,5 +1,8 @@
 drop table if exists pet CASCADE;
 drop table if exists pet_type CASCADE;
+drop table if exists vet CASCADE;
+
+create table vet (id integer auto_increment, first_name varchar(255) not null, last_name varchar(255) not null, primary key (id));
 
 create table pet_type (id integer, name varchar(255) not null, primary key (id));
 create table pet (id integer auto_increment, name varchar(255) not null, type_id integer not null, primary key (id));